frontend-otel/util/requirements.txt
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
protobuf==4.24.3
grpcio==1.58.0
